// src/app/shared/services/model/model.ts
export interface pageSelection {
  skip: number;
  limit: number;
}
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
  export interface userReview {
    id: number;
    carName: string;
    rentalType: string;
    review: string;
    ratings: string;
    img: string;
    feature: string;
    isSelected: boolean;
    rating: string;
  }